Change the position of the profile picture
I used a transform property here to move the image back half of its width and half of its height to center it properly.

Then I used the height of the top background image to position the center of the profile pic at the bottom of the top background image.